% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul-response.R
\name{simul_response}
\alias{simul_response}
\title{Simulate spectral sensor response}
\usage{
simul_response(
  source.spct,
  sensor.mspct,
  range = NULL,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  quantity = "total",
  time.unit = NULL,
  scale.factor = 1
)
}
\arguments{
\item{source.spct}{source_spct object Light source spectral irradiance.}

\item{range}{numeric vector of length two Range of wavelengths to use
(nanometres, nm)}

\item{unit.out}{character Allowed values \code{"energy"}, and
\code{"photon"}, or its alias \code{"quantum"}.}

\item{quantity}{character string One of \code{"average"} or \code{"mean"},
\code{"total"}, \code{"contribution"}, \code{"contribution.pc"},
\code{"relative"} or \code{"relative.pc"}.}

\item{time.unit}{character or lubridate::duration object.}

\item{scale.factor}{numeric vector of length 1, or length equal to that of
\code{w.band}. Numeric multiplier applied to returned values.}

\item{responce.spct}{or response.mspct Light sensor spectral responsiveness,
one or more channels.}
}
\description{
Simulate spectral sensor response
}
\details{
Compute sensor response spectrum by convolution of light source
emission spectrum or illumination spectrum and the responsiveness spectrum
of a sensor with one or more channels. Return the integral over wavelengths
for each sensor channel.
}
\examples{
simul_response(sun.spct, ams_AS7343.spct)
simul_response(sun.spct, ccd.spct)
simul_response(sun.spct, ccd.spct, unit.out = "photon")

}
